internal/agent/agent_test.go
package agent_test

import (
	"context"
	"crypto/tls"
	"fmt"
	"io/ioutil"
	"os"
	"testing"
	"time"

	api "github.com/abdulmajid18/log-distributed-system/api/v1"
	"github.com/abdulmajid18/log-distributed-system/internal/agent"
	"github.com/abdulmajid18/log-distributed-system/internal/config"
	"github.com/stretchr/testify/require"
	"github.com/travisjeffery/go-dynaport"
	"google.golang.org/grpc"
	"google.golang.org/grpc/credentials"
)

func Test(t *testing.T) {
	serverTLSConfig, err := config.SetupTLSConfig(config.TLSConfig{
		CertFile:      config.ServerCertFile,
		KeyFile:       config.ServerKeyFile,
		CAFile:        config.CAFile,
		Server:        true,
		ServerAddress: "127.0.0.1",
	})

	require.NoError(t, err)
	peerTLSConfig, err := config.SetupTLSConfig(config.TLSConfig{
		CertFile:      config.RootClientCertFile,
		KeyFile:       config.RootClientKeyFile,
		CAFile:        config.CAFile,
		Server:        false,
		ServerAddress: "127.0.0.1",
	})
	require.NoError(t, err)

	var agents []*agent.Agent

	for i := 0; i < 3; i++ {
		ports := dynaport.Get(2)
		bindAddr := fmt.Sprintf("%s:%d", "127.0.0.1", ports[0])
		rpcPort := ports[1]

		dataDir, _ := ioutil.TempDir("", "agent-test-log")

		var startJoinAddrs []string

		if i != 0 {
			startJoinAddrs = append(startJoinAddrs, agents[0].Config.BindAddr)
		}

		agent, err := agent.New(agent.Config{
			NodeName:        fmt.Sprintf("%d", i),
			BindAddr:        bindAddr,
			PeerTLSConfig:   peerTLSConfig,
			ServerTLSConfig: serverTLSConfig,
			DataDir:         dataDir,
			RPCPort:         rpcPort,
			StartJoinAddrs:  startJoinAddrs,
			ACLModelFile:    config.ACLModelFile,
			ACLModelPolicy:  config.ACLPolicyFile,
		})

		require.NoError(t, err)
		agents = append(agents, agent)

	}

	defer func() {
		for _, agent := range agents {
			err := agent.Shutdown()
			require.NoError(t, err)
			require.NoError(t, os.Remove(agent.Config.DataDir))
		}
	}()

	time.Sleep(3 * time.Second)

	//Single node produce and consume
	leaderClient := client(t, agents[0], peerTLSConfig)
	produceResponse, err := leaderClient.Produce(context.Background(), &api.ProduceRequest{
		Record: &api.Record{
			Value: []byte("chicken wings"),
		},
	})

	require.NoError(t, err)
	consumeResponse, err := leaderClient.Consume(context.Background(), &api.ConsumeRequest{
		Offset: 0,
	})
	require.NoError(t, err)
	require.Equal(t, consumeResponse.Record.Value, []byte("chicken wings"))

	time.Sleep(3 * time.Second)

	followerClient := client(t, agents[1], peerTLSConfig)

	consumeResponse, err = followerClient.Consume(context.Background(),
		&api.ConsumeRequest{Offset: produceResponse.Offset})
	require.NoError(t, err)
	require.Equal(t, consumeResponse.Record.Value, []byte("chicken wings"))
}

func client(t *testing.T, agent *agent.Agent, tlsConfig *tls.Config) api.LogClient {
	tlsCreds := credentials.NewTLS(tlsConfig)
	opts := []grpc.DialOption{grpc.WithTransportCredentials(tlsCreds)}
	rpcAddr, err := agent.Config.RPCAddr()
	require.NoError(t, err)
	conn, err := grpc.Dial(fmt.Sprintf(
		"%s",
		rpcAddr,
	), opts...)
	require.NoError(t, err)
	client := api.NewLogClient(conn)
	return client
}
